Departments radiation functions are now main functions
