+++ test/test_trip_sampler.py
@@ -1,8 +1,6 @@
 import mobility
 from mobility.carbon_computation import carbon_computation
 
-import pytest
-
 
 def test_trip_sampler():
     # Create trip sampler
@@ -18,8 +16,9 @@ def test_trip_sampler():
     )
     # Compute carbon emissions for each trip
     emissions = carbon_computation(trips, ademe_database="Base_Carbone_V22.0.csv")
-    assert emissions["carbon_emissions"].sum() < 50000
-    assert emissions["carbon_emissions"].sum() > 100
+    total_emissions = emissions["carbon_emissions"].sum()
+    assert total_emissions < 50000
+    assert total_emissions > 100
 
     ts2 = mobility.TripSampler()
     trips2 = ts2.get_trips(
